+++ targets.h
@@ -120,15 +120,15 @@ int sendpingquery(int sd, int rawsd, Target *target,
 		  u16 seq, unsigned short id, struct scanstats *ss, 
 		  struct timeval *time, int pingtype, struct pingtech ptech);
 int sendrawtcpudppingqueries(int rawsd, Target *target, int pingtype,
-			  int seq, struct timeval *time, struct pingtune *pt);
+			  u16 seq, struct timeval *time, struct pingtune *pt);
 int sendrawtcppingquery(int rawsd, Target *target, int pingtype, u16 probe_port,
-			int seq, struct timeval *time, struct pingtune *pt);
+			u16 seq, struct timeval *time, struct pingtune *pt);
 int sendrawudppingquery(int rawsd, Target *target, u16 probe_port,
-			int seq, struct timeval *time, struct pingtune *pt);
+			u16 seq, struct timeval *time, struct pingtune *pt);
 int sendconnecttcpqueries(Target *hostbatch[], struct tcpqueryinfo *tqi, Target *target,
-			  int seq, struct timeval *time, struct pingtune *pt, struct timeout_info *to, int max_width);
+			  u16 seq, struct timeval *time, struct pingtune *pt, struct timeout_info *to, int max_width);
 int sendconnecttcpquery(Target *hostbatch[], struct tcpqueryinfo *tqi, Target *target, int probe_port_num,
-			int seq, struct timeval *time, struct pingtune *pt, struct timeout_info *to, int max_width);
+			u16 seq, struct timeval *time, struct pingtune *pt, struct timeout_info *to, int max_width);
 int get_connecttcpscan_results(struct tcpqueryinfo *tqi, 
 			       Target *hostbatch[], 
 			       struct timeval *time, struct pingtune *pt,
